opendal_service_lakefs

Struct Lakefs

pub struct Lakefs { /* private fields */ }

Lakefs’s API support. This service will visit the Lakefs API to access the Lakefs File System. Currently, we only support the model and dataset types of repositories, and operations are limited to reading and listing/stating.

Lakefs doesn’t host official HTTP API docs. Detailed HTTP request API information can be found on the Lakefs_hub Source Code.

Capabilities

Depending on its configuration and the backing system, this service can expose:

create_dir

stat

read

write

delete

list

copy

rename

presign

Inspect the effective capability set with [opendal_core::Operator::info] and [opendal_core::OperatorInfo::capability] after building an operator.

Configuration

Use crate::LakefsConfig for serializable configuration and this builder’s methods for direct construction. The field and method documentation defines accepted values, defaults, and environment interaction.

Examples

Via Builder

use opendal_core::Operator;
use opendal_core::Result;
use opendal_service_lakefs::Lakefs;

#[tokio::main]
async fn main() -> Result<()> {
    // Create Lakefs backend builder
    let mut builder = Lakefs::default()
        // set the type of Lakefs endpoint
        .endpoint("https://whole-llama-mh6mux.us-east-1.lakefscloud.io")
        // set the id of Lakefs repository
        .repository("sample-repo")
        // set the branch of Lakefs repository
        .branch("main")
        // set the username for accessing the repository
        .username("xxx")
        // set the password for accessing the repository
        .password("xxx");

    let op: Operator = Operator::new(builder)?;

    let stat = op.stat("README.md").await?;
    println!("{:?}", stat);
    Ok(())
}

Implementations

impl LakefsBuilder

pub fn endpoint(self, endpoint: &str) -> Self

Set the endpoint of this backend.

endpoint must be full uri.

This is required.

http://127.0.0.1:8000 (lakefs daemon in local)

https://my-lakefs.example.com (lakefs server)

pub fn username(self, username: &str) -> Self

Set username of this backend. This is required.

pub fn password(self, password: &str) -> Self

Set password of this backend. This is required.

pub fn branch(self, branch: &str) -> Self

Set branch of this backend or a commit ID. Default is main.

Branch can be a branch name.

For example, branch can be:

main

1d0c4eb

pub fn root(self, root: &str) -> Self

Set root of this backend.

All operations will happen under this root.

pub fn repository(self, repository: &str) -> Self

Set the repository of this backend.

This is required.